requirements.txt
python-telegram-bot
requests
flask[async]
pyyaml
